Trying to consolidate week answer solution creation into one struct
